# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DistriMed - Catálogo de Productos Médicos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  vendedor: "vendedor",
  viewer: "viewer",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  email?: string;
  role: UserRole;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  vendedor: "vendedor",
  viewer: "viewer",
} as const;

export interface CreateUserRequest {
  username: string;
  name: string;
  email?: string;
  password: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  vendedor: "vendedor",
  viewer: "viewer",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  password?: string;
  role?: UpdateUserRequestRole;
}

export interface Product {
  id: number;
  name: string;
  code: string;
  description?: string;
  categoryId: number;
  categoryName?: string;
  price: number;
  stock: number;
  image?: string;
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateProductRequest {
  name: string;
  code: string;
  description?: string;
  categoryId: number;
  price: number;
  stock: number;
  image?: string;
}

export interface UpdateProductRequest {
  name?: string;
  code?: string;
  description?: string;
  categoryId?: number;
  price?: number;
  stock?: number;
  image?: string;
}

export interface Category {
  id: number;
  name: string;
  description?: string;
  productCount?: number;
  createdAt: string;
}

export interface CreateCategoryRequest {
  name: string;
  description?: string;
}

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  totalUsers: number;
  lowStockProducts: number;
  totalStockValue: number;
}

export interface CategoryCount {
  categoryName: string;
  count: number;
}

export type GetProductsParams = {
  search?: string;
  category?: string;
  page?: number;
  limit?: number;
};