# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DistriMed - Catálogo de Productos Médicos API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary User login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    name: zod.string(),
    email: zod.string().optional(),
    role: zod.enum(["admin", "vendedor", "viewer"]),
    createdAt: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary User logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  email: zod.string().optional(),
  role: zod.enum(["admin", "vendedor", "viewer"]),
  createdAt: zod.string(),
});

/**
 * @summary Get all products
 */
export const GetProductsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
  page: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const GetProductsResponse = zod.object({
  products: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      code: zod.string(),
      description: zod.string().optional(),
      categoryId: zod.number(),
      categoryName: zod.string().optional(),
      price: zod.number(),
      stock: zod.number(),
      image: zod.string().optional(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  code: zod.string(),
  description: zod.string().optional(),
  categoryId: zod.number(),
  price: zod.number(),
  stock: zod.number(),
  image: zod.string().optional(),
});

/**
 * @summary Get product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  code: zod.string(),
  description: zod.string().optional(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  price: zod.number(),
  stock: zod.number(),
  image: zod.string().optional(),
  createdAt: zod.string(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  code: zod.string().optional(),
  description: zod.string().optional(),
  categoryId: zod.number().optional(),
  price: zod.number().optional(),
  stock: zod.number().optional(),
  image: zod.string().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  code: zod.string(),
  description: zod.string().optional(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  price: zod.number(),
  stock: zod.number(),
  image: zod.string().optional(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteProductResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all categories
 */
export const GetCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  productCount: zod.number().optional(),
  createdAt: zod.string(),
});
export const GetCategoriesResponse = zod.array(GetCategoriesResponseItem);

/**
 * @summary Create a category
 */
export const CreateCategoryBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Get all users (admin only)
 */
export const GetUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  email: zod.string().optional(),
  role: zod.enum(["admin", "vendedor", "viewer"]),
  createdAt: zod.string(),
});
export const GetUsersResponse = zod.array(GetUsersResponseItem);

/**
 * @summary Create a user (admin only)
 */
export const CreateUserBody = zod.object({
  username: zod.string(),
  name: zod.string(),
  email: zod.string().optional(),
  password: zod.string(),
  role: zod.enum(["admin", "vendedor", "viewer"]),
});

/**
 * @summary Update a user
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  password: zod.string().optional(),
  role: zod.enum(["admin", "vendedor", "viewer"]).optional(),
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  email: zod.string().optional(),
  role: zod.enum(["admin", "vendedor", "viewer"]),
  createdAt: zod.string(),
});

/**
 * @summary Delete a user
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteUserResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalProducts: zod.number(),
  totalCategories: zod.number(),
  totalUsers: zod.number(),
  lowStockProducts: zod.number(),
  totalStockValue: zod.number(),
});

/**
 * @summary Get product count by category
 */
export const GetProductsByCategoryResponseItem = zod.object({
  categoryName: zod.string(),
  count: zod.number(),
});
export const GetProductsByCategoryResponse = zod.array(
  GetProductsByCategoryResponseItem,
);

/**
 * @summary Get recently added products
 */
export const GetRecentProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  code: zod.string(),
  description: zod.string().optional(),
  categoryId: zod.number(),
  categoryName: zod.string().optional(),
  price: zod.number(),
  stock: zod.number(),
  image: zod.string().optional(),
  createdAt: zod.string(),
});
export const GetRecentProductsResponse = zod.array(
  GetRecentProductsResponseItem,
);