# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DistriMed - Catálogo de Productos Médicos API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalProducts: number;
  totalCategories: number;
  totalUsers: number;
  lowStockProducts: number;
  totalStockValue: number;
}