# lib/api-zod/src/generated/types/updateProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DistriMed - Catálogo de Productos Médicos API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductRequest {
  name?: string;
  code?: string;
  description?: string;
  categoryId?: number;
  price?: number;
  stock?: number;
  image?: string;
}